refactor: reorganize session card components and simplify date formatting
